+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.FinalProjectApartTalk"
         tools:targetApi="31">
         <activity
-            android:name=".CommunityActivity"
+            android:name=".ui.community.activity.CommunityActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
