+++ app/src/main/java/kr/co/lion/application/finalproject_aparttalk/MainActivity.kt
@@ -5,13 +5,10 @@ import android.os.Handler
 import android.os.Looper
 import android.os.SystemClock
 import android.widget.Toast
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.FragmentManager
 import kr.co.lion.application.finalproject_aparttalk.databinding.ActivityMainBinding
-import kr.co.lion.application.finalproject_aparttalk.ui.community.CommunityFragment
+import kr.co.lion.application.finalproject_aparttalk.ui.community.fragment.CommunityFragment
 import kr.co.lion.application.finalproject_aparttalk.ui.entiremenu.EntireMenuFragment
 import kr.co.lion.application.finalproject_aparttalk.ui.facility.FacilityFragment
 import kr.co.lion.application.finalproject_aparttalk.ui.home.AlarmFragment
