+++ app/src/main/java/kr/co/lion/application/finalproject_aparttalk/ui/community/fragment/TabTradeFragment.kt
@@ -1,4 +1,4 @@
-package kr.co.lion.application.finalproject_aparttalk.ui.community
+package kr.co.lion.application.finalproject_aparttalk.ui.community.fragment
 
 import android.content.Intent
 import android.os.Bundle
@@ -9,12 +9,11 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.divider.MaterialDividerItemDecoration
-import kr.co.lion.application.finalproject_aparttalk.CommunityActivity
+import kr.co.lion.application.finalproject_aparttalk.ui.community.activity.CommunityActivity
 import kr.co.lion.application.finalproject_aparttalk.MainActivity
-import kr.co.lion.application.finalproject_aparttalk.R
 import kr.co.lion.application.finalproject_aparttalk.databinding.FragmentTabTradeBinding
-import kr.co.lion.application.finalproject_aparttalk.databinding.RowCommunityTabQuestionBinding
 import kr.co.lion.application.finalproject_aparttalk.databinding.RowCommunityTabTradeBinding
+import kr.co.lion.application.finalproject_aparttalk.ui.community.adapter.CommunityTabTradeRecyclerViewAdapter
 import kr.co.lion.application.finalproject_aparttalk.util.CommunityFragmentName
 
 class TabTradeFragment : Fragment() {
@@ -71,58 +70,10 @@ class TabTradeFragment : Fragment() {
     private fun settingRecyclerViewTabTrade() {
         fragmentTabTradeBinding.apply {
             recyclerViewTabTrade.apply {
-                adapter = CommunityTabTradeRecyclerViewAdapter()
+                adapter = CommunityTabTradeRecyclerViewAdapter(requireContext())
                 layoutManager = LinearLayoutManager(mainActivity)
                 val deco = MaterialDividerItemDecoration(mainActivity, MaterialDividerItemDecoration.VERTICAL)
                 addItemDecoration(deco)
-            }
-        }
-    }
-
-    // 커뮤니티 거래 탭 리사이클러뷰 어댑터
-    inner class CommunityTabTradeRecyclerViewAdapter : RecyclerView.Adapter<CommunityTabTradeRecyclerViewAdapter.CommunityTabTradeViewHolder>() {
-        inner class CommunityTabTradeViewHolder(rowCommunityTabTradeBinding: RowCommunityTabTradeBinding) : RecyclerView.ViewHolder(rowCommunityTabTradeBinding.root) {
-            val rowCommunityTabTradeBinding : RowCommunityTabTradeBinding
-
-            init {
-                this.rowCommunityTabTradeBinding = rowCommunityTabTradeBinding
-
-                this.rowCommunityTabTradeBinding.root.layoutParams = ViewGroup.LayoutParams(
-                    ViewGroup.LayoutParams.MATCH_PARENT,
-                    ViewGroup.LayoutParams.WRAP_CONTENT
-                )
-            }
-
-        }
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CommunityTabTradeViewHolder {
-            val rowCommunityTabTradeBinding = RowCommunityTabTradeBinding.inflate(layoutInflater)
-            val communityTabTradeViewHolder = CommunityTabTradeViewHolder(rowCommunityTabTradeBinding)
-
-            return communityTabTradeViewHolder
-        }
-
-        override fun getItemCount(): Int {
-            return 10
-        }
-
-        override fun onBindViewHolder(holder: CommunityTabTradeViewHolder, position: Int) {
-            holder.rowCommunityTabTradeBinding.apply {
-                textViewCommunityListLabelTrade.text = "거래"
-                textViewCommunityListTitleTrade.text = "글 제목입니다 $position"
-                textViewCommunityListContentTrade.text = "글 내용입니다 글 내용입니다 글 내용입니다\n" +
-                        "글 내용입니다 글 내용입니다 글 내용입니다 "
-                textViewCommunityListLikeCntTrade.text = "999"
-                textViewCommunityListCommentCntTrade.text = "999"
-                textViewCommunityListDateTrade.text = "2024.05.17"
-
-                linearLayoutCommunityListTrade.setOnClickListener {
-                    val intent = Intent(mainActivity, CommunityActivity::class.java)
-                    intent.putExtra("fragmentName", CommunityFragmentName.COMMUNITY_DETAIL_FRAGMENT)
-                    // 게시글 번호도 주기
-                    // communityIntent.putExtra("postIdx", searchList[position].postIdx)
-                    startActivity(intent)
-                }
             }
         }
     }
