+++ app/src/main/res/layout/fragment_community.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.community.CommunityFragment"
+    tools:context=".ui.community.fragment.CommunityFragment"
     android:background="@color/white">
 
     <com.google.android.material.appbar.MaterialToolbar
