+++ app/src/main/res/layout/fragment_community_bottom_sheet.xml
@@ -6,7 +6,7 @@
     android:background="@drawable/bottom_sheet_stroke"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".ui.community.CommunityBottomSheetFragment">
+    tools:context=".ui.community.fragment.CommunityBottomSheetFragment">
 
     <View
         android:id="@+id/divider5"
