+++ app/src/main/res/layout/fragment_community_detail.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/white"
-    tools:context=".ui.community.CommunityDetailFragment">
+    tools:context=".ui.community.fragment.CommunityDetailFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbarCommunityDetail"
