+++ app/src/main/res/layout/fragment_community_modify.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.community.CommunityModifyFragment"
+    tools:context=".ui.community.fragment.CommunityModifyFragment"
     android:background="@color/white">
 
     <com.google.android.material.appbar.MaterialToolbar
