+++ app/src/main/res/layout/fragment_community_search.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/white"
-    tools:context=".ui.community.CommunitySearchFragment">
+    tools:context=".ui.community.fragment.CommunitySearchFragment">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/textInputLayoutCommunitySearch"
