+++ app/src/main/res/layout/fragment_tab_trade.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.community.TabNotificationFragment"
+    tools:context=".ui.community.fragment.TabNotificationFragment"
     android:background="@color/white">
